# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Storyteller API","version":"0.1.0"},"paths":{"/api/health":{"get":{"summary":"Health Check","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthCheckResponse"}}}}}}},"/api/stories/":{"get":{"tags":["stories"],"summary":"List Stories","operationId":"list_stories_api_stories__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Story"},"type":"array","title":"Response List Stories Api Stories  Get"}}}}}},"post":{"tags":["stories"],"summary":"Create Story Endpoint","operationId":"create_story_endpoint_api_stories__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPreferences"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Story"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/stories/{story_id}":{"get":{"tags":["stories"],"summary":"Get Story By Id","operationId":"get_story_by_id_api_stories__story_id__get","parameters":[{"name":"story_id","in":"path","required":true,"schema":{"type":"string","title":"Story Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Story"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["stories"],"summary":"Delete Story By Id","operationId":"delete_story_by_id_api_stories__story_id__delete","parameters":[{"name":"story_id","in":"path","required":true,"schema":{"type":"string","title":"Story Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Chapter":{"properties":{"content":{"type":"string","title":"Content","description":"El contenido detallado del capítulo"},"image_description":{"type":"string","title":"Image Description","description":"\nDescripción detallada para la generación de una imagen clave del capítulo, \nla descripción debe ser en inglés y debe ser lo más detallada posible\n        "},"id":{"type":"integer","title":"Id","description":"El ID del capítulo"},"title":{"type":"string","title":"Title","description":"El título del capítulo"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url","description":"La URL de la imagen del capítulo"}},"type":"object","required":["content","image_description","id","title","image_url"],"title":"Chapter"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthCheckResponse":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"HealthCheckResponse"},"Story":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"title":{"type":"string","title":"Title"},"image_url":{"type":"string","title":"Image Url"},"premise":{"type":"string","title":"Premise"},"chapters":{"items":{"$ref":"#/components/schemas/Chapter"},"type":"array","title":"Chapters"},"preferences":{"$ref":"#/components/schemas/UserPreferences"}},"type":"object","required":["title","image_url","premise","chapters","preferences"],"title":"Story"},"UserPreferences":{"properties":{"genre":{"type":"string","title":"Genre","description":"El género literario de la historia"},"length":{"type":"string","title":"Length","description":"La longitud deseada de la historia (corta, media, larga)"},"style":{"type":"string","title":"Style","description":"El estilo de escritura deseado"},"themes_to_include":{"items":{"type":"string"},"type":"array","title":"Themes To Include","description":"Lista de temas que deben incluirse en la historia"},"themes_to_avoid":{"items":{"type":"string"},"type":"array","title":"Themes To Avoid","description":"Lista de temas que deben evitarse en la historia"},"art_style":{"type":"string","title":"Art Style","description":"El estilo artístico deseado para las imágenes"}},"type":"object","required":["genre","length","style","themes_to_include","themes_to_avoid","art_style"],"title":"UserPreferences"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}